Use courseProgress for progress and fix profile layouts
Replace usages of the level-based progress value with courseProgress in HomeTab and ProgressSheet so the displayed progress matches course completion. Remove the now-unused _progressForXp helper. Adjust ProfileTab layout to avoid overflow by wrapping text/columns with Expanded, adding spacing (SizedBox), and limiting xp-needed text to 2 lines with ellipsis. Also fix a type issue in student_home_screen by casting computed height to double. Changes applied across main_screen.dart and student_home_screen.dart.

+++ mandarinmate/lib/screens/student_home_screen.dart
@@ -263,7 +263,7 @@ class _StudentHomeScreenState extends State<StudentHomeScreen> {
                         curve: Curves.easeOut,
                         width: 24,
                         // 80 is the max physical pixel height of the bar
-                        height: 80 * percentage,
+                        height: (80 * percentage).toDouble(),
                         decoration: BoxDecoration(
                           color: const Color(0xFFFFCC00), // DHL Yellow
                           borderRadius: BorderRadius.circular(4),
